requirements.txt
# Discord
discord.py>=2.4.0

# HTTP
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0
